deploy_fastapi/streamlit/requirements.txt
altair==5.1.2 
attrs==23.1.0 
blinker==1.7.0 
cachetools==5.3.2 
certifi==2023.7.22 
charset-normalizer==3.3.2 
click==8.1.7 
colorama==0.4.6 
gitdb==4.0.11 
gitpython==3.1.40 
idna==3.4 
importlib-metadata==6.8.0 
jinja2==3.1.2 
jsonschema-specifications==2023.7.1 
jsonschema==4.19.2 
markdown-it-py==3.0.0 
markupsafe==2.1.3 
mdurl==0.1.2 
numpy==1.26.2 
opencv-python==4.8.1.78 
packaging==23.2 
pandas==2.1.3 
pillow==10.1.0 
protobuf==4.25.0 
pyarrow==14.0.1 
pydeck==0.8.0 
pygments==2.16.1 
python-dateutil==2.8.2 
pytz==2023.3.post1 
referencing==0.30.2 
requests-toolbelt==1.0.0 
requests==2.31.0 
rich==13.6.0 
rpds-py==0.12.0 
six==1.16.0 
smmap==5.0.1 
streamlit-image-comparison==0.0.4 
streamlit==1.28.1 
tenacity==8.2.3 
toml==0.10.2 
toolz==0.12.0 
tornado==6.3.3 
typing-extensions==4.8.0 
tzdata==2023.3 
tzlocal==5.2 
urllib3==2.1.0 
validators==0.22.0 
watchdog==3.0.0 
zipp==3.17.0 
